:sparkles: Remove reflection-based class model when validating.
Separation of the validation from the _effect_ caused by inputs not being
valid was necessary in order to make the validation work correctly in the
web environment. This uses more reasonable inheritance.

On a side note, this also does away with default in-DOM error messages, as
the logic to support these was too complex.

Refs #2.