:sparkles: Add method on FieldValidator to determine if elements match selector.
We need to be able to reliably determine if subclasses of FieldValidator
match given elements in order to be able to delegate to a given class a
message coming from higher up in the messaging chain.

This will enable us to delegate calls to ComponentInstances to individual
FieldValidator (or any class, really).

Refs #16.