:bug: Mitigate issue where flashing of messages can happen on first render.
Sometimes when initializing ComponentBehaviors, we have to wait until the
DOM has been loaded. If we can wait 100ms BEFORE making a message
invisible, it's highly likely that we'll render at least one frame before
we set the CSS attributes for message to invisible. This reduces the cycle
time between successive checks for whether the DOM is loaded to help
alleviate this problem.

The only sure-fire way to mitigate this problem is to set the default CSS
for PREFIX_formFieldMessage to 'visibility: hidden', so it isn't rendered,
even before ComponentBehaviors is initialized.

Fixes #18.